reorganize; add test
